scripts/research/mesh/legacy/v5/requirements.txt
aiohttp
pyyaml
matplotlib
numpy
sqlite3
